Crate ic_registry_transport

Modules

pb

Enums

Error
The possible errors in registry responses. Per key errors are associated with a particular key.

Functions

delete
Shorthand to create a RegistryMutation with type Delete.

deserialize_atomic_mutate_request
Deserializes the arguments for a request to the atomic_mutate() function in the registry canister, from protobuf.

deserialize_atomic_mutate_response
Deserializes the response obtained from the registry canister for a get() call from protobuf.

deserialize_get_changes_since_request
Deserializes the response obtained from the registry canister for a get_changes_since() call, from protobuf.

deserialize_get_changes_since_response
Deserializes the response obtained from the registry canister for a get_changes_since() call, from protobuf.

deserialize_get_latest_version_response
Deserializes the response obtained from the registry canister for a get_latest_version() call, from protobuf.

deserialize_get_value_request
Deserializes the arguments for a request to the get_value() function in the registry canister, from protobuf.

deserialize_get_value_response
Deserializes the response obtained from the registry canister for a get_value() call, from protobuf.

insert
Shorthand to create a RegistryMutation with type Insert.

precondition
Shorthand to create a Precondition.

serialize_atomic_mutate_request
Serializes the arguments for a request to the insert() function in the registry canister, into protobuf.

serialize_atomic_mutate_response
Serializes a response for a atomic_mutate() request to the registry canister.

serialize_get_changes_since_request
Serializes a request for a get_changes_since() request to the registry canister.

serialize_get_changes_since_response
Serializes a response for a get_changes_since() request to the registry canister.

serialize_get_latest_version_response
Serializes a response for a get_latest_version() request to the registry canister.

serialize_get_value_request
Serializes the arguments for a request to the get_value() function in the registry canister, into protobuf.

serialize_get_value_response
Serializes a response for a get_value() request to the registry canister.

update
Shorthand to create a RegistryMutation with type Update.

upsert
Shorthand to create a RegistryMutation with type Upsert.